requirements.txt
streamlit>=1.40
google-genai>=1.0
PyMuPDF>=1.24
python-dotenv>=1.0
requests>=2.28
